Further clean up engine/mod, use thiserror
