+++ Cargo.toml
@@ -19,6 +19,7 @@ parking_lot = { version = "0.12.4", features = ["send_guard"] }
 fxhash = "0.2.1"
 static_assertions = "1.1.0"
 rayon = "1.10.0"
+thiserror = "2"
 
 [dev-dependencies]
 criterion = "0.6.0"
